[OBEY-CAMPAIGN-8deed8b4] refactor: split oversized init and helper files
Keep PR-touched source and test files under the 500-line target by splitting init command helpers, scaffold init tests, and integration container helpers by responsibility.

Verification:
- go test ./cmd/camp ./internal/scaffold
- go test -count=1 -tags integration ./tests/integration -run 'TestCampInit|TestCampCreate|TestCampSettings'\n- golangci-lint run --new-from-rev HEAD ./cmd/camp ./internal/scaffold\n- golangci-lint run --new-from-rev HEAD --build-tags integration ./tests/integration
