+++ cmd/camp/init.go
@@ -12,13 +12,9 @@ import (
 	"github.com/Obedience-Corp/camp/internal/campaign"
 	"github.com/Obedience-Corp/camp/internal/config"
 	camperrors "github.com/Obedience-Corp/camp/internal/errors"
-	"github.com/Obedience-Corp/camp/internal/fest"
-	"github.com/Obedience-Corp/camp/internal/git/commit"
 	"github.com/Obedience-Corp/camp/internal/nav/tui"
 	"github.com/Obedience-Corp/camp/internal/scaffold"
 	"github.com/Obedience-Corp/camp/internal/ui"
-	"github.com/Obedience-Corp/camp/internal/ui/theme"
-	"github.com/charmbracelet/huh"
 	"github.com/spf13/cobra"
 )
 
@@ -342,402 +338,4 @@ func runInitFlow(ctx context.Context, p initParams, w initWriters, isInteractive
 	}
 
 	return nil
-}
-
-// initializeFestivals runs fest init in the campaign directory.
-// Returns true if successful, false with guidance if fest is unavailable.
-func initializeFestivals(ctx context.Context, campaignRoot string, w initWriters) (bool, error) {
-	// Check if already initialized
-	if fest.IsInitialized(campaignRoot) {
-		writeLine(w.humanOut, ui.Success("Festival Methodology already initialized"))
-		return true, nil
-	}
-
-	// Check if fest CLI is available
-	if !fest.IsFestAvailable() {
-		showFestInstallGuidance(w)
-		return false, fest.ErrFestNotFound
-	}
-
-	// Check if festivals directory has content but isn't initialized
-	if hasNonFestContent(campaignRoot) {
-		showFestManualInitGuidance(w)
-		return false, nil
-	}
-
-	writeLine(w.humanOut, ui.Info("Initializing Festival Methodology..."))
-	err := fest.RunInit(ctx, &fest.InitOptions{
-		CampaignRoot: campaignRoot,
-	})
-	if err != nil {
-		showFestInitFailure(err, w)
-		return false, err
-	}
-
-	writeLine(w.humanOut, ui.Success("Festival Methodology initialized"))
-	return true, nil
-}
-
-// hasNonFestContent checks if festivals/ has content that isn't fest-initialized.
-func hasNonFestContent(campaignRoot string) bool {
-	festivalsDir := filepath.Join(campaignRoot, "festivals")
-	entries, err := os.ReadDir(festivalsDir)
-	if err != nil {
-		return false
-	}
-	// If we have entries but fest isn't initialized, there's non-fest content
-	return len(entries) > 0 && !fest.IsInitialized(campaignRoot)
-}
-
-// showFestInstallGuidance displays guidance for installing fest CLI.
-func showFestInstallGuidance(w initWriters) {
-	writeLine(w.humanOut)
-	writeLine(w.humanOut, ui.Dim("Festival Methodology provides structured project planning."))
-	writeLine(w.humanOut, ui.Dim("Install the fest CLI to enable it:"))
-	writeLine(w.humanOut)
-	writeLine(w.humanOut, ui.Dim("  go install github.com/Obedience-Corp/fest/cmd/fest@latest"))
-	writeLine(w.humanOut)
-	writeLine(w.humanOut, ui.Dim("Then run: camp init --repair"))
-	writeLine(w.humanOut, ui.Dim("Continuing without Festival Methodology..."))
-}
-
-// showFestManualInitGuidance displays guidance when festivals/ has non-fest content.
-func showFestManualInitGuidance(w initWriters) {
-	writeLine(w.humanOut)
-	writeLine(w.humanOut, ui.Warning("festivals/ has content but is not fest-initialized"))
-	writeLine(w.humanOut, ui.Dim("Run 'fest init' manually to initialize, or clear the directory."))
-}
-
-// showFestInitFailure displays guidance when fest init fails.
-func showFestInitFailure(err error, w initWriters) {
-	writeLine(w.humanOut, ui.Warning(fmt.Sprintf("Failed to initialize Festival Methodology: %v", err)))
-	writeLine(w.humanOut)
-	writeLine(w.humanOut, ui.Dim("You may need to run 'fest init' manually."))
-	writeLine(w.humanOut, ui.Dim("Use 'fest init --help' for options."))
-	writeLine(w.humanOut, ui.Dim("Continuing with campaign creation..."))
-}
-
-// checkDirectoryEmpty verifies the target directory is empty or gets user approval.
-func checkDirectoryEmpty(dir string, force, isInteractive bool, w initWriters) error {
-	// Resolve to absolute path
-	absDir, err := filepath.Abs(dir)
-	if err != nil {
-		return camperrors.Wrap(err, "failed to resolve directory path")
-	}
-
-	// Check if directory exists
-	info, err := os.Stat(absDir)
-	if os.IsNotExist(err) {
-		// Directory doesn't exist - will be created, so it's "empty"
-		return nil
-	}
-	if err != nil {
-		return camperrors.Wrap(err, "failed to check directory")
-	}
-	if !info.IsDir() {
-		return fmt.Errorf("path exists but is not a directory: %s", absDir)
-	}
-
-	// Check if directory has contents
-	entries, err := os.ReadDir(absDir)
-	if err != nil {
-		return camperrors.Wrap(err, "failed to read directory")
-	}
-
-	if len(entries) == 0 {
-		// Directory is empty
-		return nil
-	}
-
-	// Directory is not empty
-	if force {
-		// User explicitly approved via --force flag
-		return nil
-	}
-
-	if isInteractive {
-		// Prompt for confirmation. The prompt text goes to humanOut (stderr in
-		// print-path mode), which is the conventional channel for interactive prompts.
-		writeLine(w.humanOut, ui.Warning(fmt.Sprintf("Directory '%s' is not empty.", filepath.Base(absDir))))
-		write(w.humanOut, "Continue and initialize campaign here? [y/N]: ")
-
-		reader := bufio.NewReader(os.Stdin)
-		response, err := reader.ReadString('\n')
-		if err != nil {
-			return camperrors.Wrap(err, "failed to read response")
-		}
-
-		response = strings.TrimSpace(strings.ToLower(response))
-		if response != "y" && response != "yes" {
-			return fmt.Errorf("initialization cancelled")
-		}
-		writeLine(w.humanOut)
-		return nil
-	}
-
-	// Non-interactive mode without --force
-	return fmt.Errorf("directory '%s' is not empty\n       Use --force to initialize anyway, or run in an interactive terminal to confirm", filepath.Base(absDir))
-}
-
-// collectCampaignInfo gathers description and mission from user input.
-// In interactive mode, prompts via TUI. In non-interactive mode, requires flags.
-func collectCampaignInfo(ctx context.Context, description, mission string, isInteractive bool) (string, string, error) {
-	// If both are provided via flags, use them
-	if description != "" && mission != "" {
-		return description, mission, nil
-	}
-
-	// Non-interactive mode without required fields
-	if !isInteractive {
-		if description == "" && mission == "" {
-			return "", "", fmt.Errorf("--description and --mission are required in non-interactive mode\n       Use -d/--description and -m/--mission flags, or run in an interactive terminal")
-		}
-		if description == "" {
-			return "", "", fmt.Errorf("--description is required in non-interactive mode\n       Use -d/--description flag, or run in an interactive terminal")
-		}
-		if mission == "" {
-			return "", "", fmt.Errorf("--mission is required in non-interactive mode\n       Use -m/--mission flag, or run in an interactive terminal")
-		}
-	}
-
-	// Interactive mode - prompt for missing values
-	form := huh.NewForm(
-		huh.NewGroup(
-			huh.NewInput().
-				Title("Campaign Description").
-				Description("A brief description of this campaign").
-				Placeholder("e.g., AI agent orchestration framework").
-				Value(&description),
-			huh.NewText().
-				Title("Mission Statement").
-				Description("What is the goal of this campaign?").
-				Placeholder("Describe the mission in detail...").
-				CharLimit(1000).
-				Value(&mission),
-		),
-	)
-
-	if err := theme.RunForm(ctx, form); err != nil {
-		if theme.IsCancelled(err) {
-			return "", "", fmt.Errorf("initialization cancelled")
-		}
-		return "", "", camperrors.Wrap(err, "failed to collect campaign info")
-	}
-
-	// Validate that user provided values
-	if description == "" {
-		return "", "", fmt.Errorf("description is required")
-	}
-	if mission == "" {
-		return "", "", fmt.Errorf("mission is required")
-	}
-
-	return description, mission, nil
-}
-
-// handleRepairMission checks for missing mission in existing campaign and prompts if needed.
-func handleRepairMission(ctx context.Context, dir string, mission string, isInteractive bool, w initWriters) (string, error) {
-	// If mission is provided via flag, use it
-	if mission != "" {
-		return mission, nil
-	}
-
-	// Load existing campaign config to check for mission
-	absDir, err := filepath.Abs(dir)
-	if err != nil {
-		return "", nil // Can't check, proceed without
-	}
-
-	cfg, err := config.LoadCampaignConfig(ctx, absDir)
-	if err != nil {
-		return "", nil // No existing config, proceed without
-	}
-
-	// If campaign already has a mission, use it
-	if cfg.Mission != "" {
-		return cfg.Mission, nil
-	}
-
-	// Campaign is missing mission
-	if isInteractive {
-		writeLine(w.humanOut, ui.Warning(fmt.Sprintf("Campaign '%s' is missing a mission statement.", cfg.Name)))
-		writeLine(w.humanOut)
-
-		form := huh.NewForm(
-			huh.NewGroup(
-				huh.NewText().
-					Title("Mission Statement").
-					Description("What is the goal of this campaign?").
-					Placeholder("Describe the mission in detail...").
-					CharLimit(1000).
-					Value(&mission),
-			),
-		)
-
-		if err := theme.RunForm(ctx, form); err != nil {
-			if theme.IsCancelled(err) {
-				// User cancelled, proceed without mission
-				writeLine(w.humanOut, ui.Dim("Skipping mission statement"))
-				return "", nil
-			}
-			return "", camperrors.Wrap(err, "failed to collect mission")
-		}
-
-		return mission, nil
-	}
-
-	// Non-interactive mode - just warn
-	writeLine(w.humanOut, ui.Warning(fmt.Sprintf("Campaign '%s' is missing a mission statement", cfg.Name)))
-	writeLine(w.humanOut, ui.Dim("         Run 'camp init --repair' in an interactive terminal to add one"))
-	writeLine(w.humanOut)
-	return "", nil
-}
-
-// commitRepairChanges creates a git commit after a successful repair.
-func commitRepairChanges(ctx context.Context, initResult *scaffold.InitResult, plan *scaffold.RepairPlan, migrationCount int, w initWriters) {
-	hasChanges := len(initResult.DirsCreated) > 0 || len(initResult.FilesCreated) > 0 || migrationCount > 0
-	if plan != nil && len(plan.IntentMigrations) > 0 {
-		hasChanges = true
-	}
-	if !hasChanges {
-		return
-	}
-
-	cfg, err := config.LoadCampaignConfig(ctx, initResult.CampaignRoot)
-	if err != nil {
-		return
-	}
-
-	description := buildRepairCommitMessage(initResult, plan, migrationCount)
-	files := buildRepairCommitFiles(initResult, plan)
-
-	result := commit.Repair(ctx, commit.RepairOptions{
-		Options: commit.Options{
-			CampaignRoot:  initResult.CampaignRoot,
-			CampaignID:    cfg.ID,
-			Files:         files,
-			SelectiveOnly: true,
-		},
-		Description: description,
-	})
-
-	if result.Committed {
-		writef(w.humanOut, "\n%s %s\n", ui.SuccessIcon(), result.Message)
-	} else if result.Message != "" {
-		writef(w.humanOut, "\n%s %s\n", ui.InfoIcon(), result.Message)
-	}
-}
-
-func buildRepairCommitFiles(initResult *scaffold.InitResult, plan *scaffold.RepairPlan) []string {
-	files := make([]string, 0, len(initResult.FilesCreated)+len(initResult.DirsCreated))
-	files = append(files, initResult.FilesCreated...)
-	files = append(files, initResult.DirsCreated...)
-
-	if plan != nil {
-		for _, m := range plan.Migrations {
-			for _, item := range m.Items {
-				files = append(files,
-					filepath.Join(m.Source, item),
-					filepath.Join(m.Dest, item),
-				)
-			}
-		}
-		for _, m := range plan.IntentMigrations {
-			for _, item := range m.Items {
-				files = append(files,
-					filepath.Join(m.Source, item),
-					filepath.Join(m.Dest, item),
-				)
-			}
-		}
-	}
-
-	return commit.NormalizeFiles(initResult.CampaignRoot, files...)
-}
-
-// buildRepairCommitMessage constructs a descriptive commit body for repair operations.
-func buildRepairCommitMessage(initResult *scaffold.InitResult, plan *scaffold.RepairPlan, migrationCount int) string {
-	var b strings.Builder
-
-	if len(initResult.DirsCreated) > 0 {
-		fmt.Fprintf(&b, "Directories created:\n")
-		for _, d := range initResult.DirsCreated {
-			fmt.Fprintf(&b, "  - %s\n", d)
-		}
-		b.WriteString("\n")
-	}
-
-	if len(initResult.FilesCreated) > 0 {
-		fmt.Fprintf(&b, "Files created:\n")
-		for _, f := range initResult.FilesCreated {
-			fmt.Fprintf(&b, "  - %s\n", f)
-		}
-		b.WriteString("\n")
-	}
-
-	if plan != nil && migrationCount > 0 {
-		fmt.Fprintf(&b, "Migrated %d item(s):\n", migrationCount)
-		for _, m := range plan.Migrations {
-			for _, item := range m.Items {
-				fmt.Fprintf(&b, "  - %s → %s\n", filepath.Join(m.Source, item), m.Dest)
-			}
-		}
-	}
-
-	if plan != nil && len(plan.IntentMigrations) > 0 {
-		if b.Len() > 0 {
-			b.WriteString("\n")
-		}
-		fmt.Fprintf(&b, "Migrated %d legacy intent item(s):\n", countMigrationItems(plan.IntentMigrations))
-		for _, m := range plan.IntentMigrations {
-			for _, item := range m.Items {
-				fmt.Fprintf(&b, "  - %s → %s\n", filepath.Join(m.Source, item), m.Dest)
-			}
-		}
-	}
-
-	return strings.TrimRight(b.String(), "\n")
-}
-
-func countMigrationItems(migrations []scaffold.MigrationAction) int {
-	total := 0
-	for _, m := range migrations {
-		total += len(m.Items)
-	}
-	return total
-}
-
-// printRepairDiff displays the proposed repair changes as a colored diff.
-func printRepairDiff(plan *scaffold.RepairPlan, w initWriters) {
-	writeLine(w.humanOut, ui.Subheader("Repair Preview"))
-	writeLine(w.humanOut)
-
-	for _, c := range plan.Changes {
-		switch c.Type {
-		case scaffold.RepairAdd:
-			writef(w.humanOut, "  %s  %s  %s\n",
-				ui.Success("+"),
-				ui.Success(c.Key),
-				ui.Dim("("+c.Description+")"),
-			)
-		case scaffold.RepairModify:
-			writef(w.humanOut, "  %s  %s  %s\n",
-				ui.Warning("~"),
-				ui.Warning(c.Key),
-				ui.Dim("("+c.Description+")"),
-			)
-		case scaffold.RepairPreserve:
-			writef(w.humanOut, "  %s  %s  %s\n",
-				ui.Dim("✓"),
-				ui.Value(c.Key),
-				ui.Dim("(user-defined, preserved)"),
-			)
-		case scaffold.RepairMigrate:
-			writef(w.humanOut, "  %s  %s  %s\n",
-				ui.Warning("→"),
-				ui.Value(c.Key),
-				ui.Dim(c.Description),
-			)
-		}
-	}
 }
