+++ internal/scaffold/init_test.go
@@ -2,15 +2,18 @@ package scaffold
 
 import (
 	"context"
+
 	"os"
-	"os/exec"
+
 	"path/filepath"
+
 	"strings"
+
 	"testing"
+
 	"time"
 
 	"github.com/Obedience-Corp/camp/internal/config"
-	"github.com/Obedience-Corp/camp/internal/intent"
 	"github.com/Obedience-Corp/camp/internal/quest"
 )
 
@@ -451,590 +454,5 @@ func TestInit_SkipsExistingDirs(t *testing.T) {
 	}
 	if !skippedMap[docsOBEY] {
 		t.Errorf("docs/OBEY.md should be in Skipped list, got: %v", result.Skipped)
-	}
-}
-
-func TestInit_GitInitialized(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "git-campaign")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	result, err := Init(ctx, campaignDir, InitOptions{Name: "git-test", NoRegister: true})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// Check that git was initialized
-	if !result.GitInitialized {
-		t.Error("GitInitialized should be true")
-	}
-
-	// Check that .git directory exists
-	gitDir := filepath.Join(campaignDir, ".git")
-	if _, err := os.Stat(gitDir); os.IsNotExist(err) {
-		t.Errorf(".git directory was not created")
-	}
-}
-
-func TestInit_SkipGitInit(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "no-git-campaign")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	result, err := Init(ctx, campaignDir, InitOptions{
-		Name:        "no-git-test",
-		SkipGitInit: true,
-		NoRegister:  true,
-	})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// Check that git was not initialized
-	if result.GitInitialized {
-		t.Error("GitInitialized should be false when SkipGitInit is true")
-	}
-
-	// Check that .git directory does not exist
-	gitDir := filepath.Join(campaignDir, ".git")
-	if _, err := os.Stat(gitDir); err == nil {
-		t.Errorf(".git directory should not exist when SkipGitInit is true")
-	}
-}
-
-func TestInit_GitAlreadyInRepo(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	// Create a git repo first
-	gitRepoDir := filepath.Join(tmpDir, "git-repo")
-	os.MkdirAll(gitRepoDir, 0755)
-
-	// Initialize git in the parent directory
-	cmd := exec.Command("git", "init")
-	cmd.Dir = gitRepoDir
-	if err := cmd.Run(); err != nil {
-		t.Skipf("git not available, skipping test: %v", err)
-	}
-
-	// Create campaign dir inside the git repo
-	campaignDir := filepath.Join(gitRepoDir, "campaign")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	result, err := Init(ctx, campaignDir, InitOptions{Name: "in-git", NoRegister: true})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// Check that git was NOT initialized (already in a repo)
-	if result.GitInitialized {
-		t.Error("GitInitialized should be false when already inside a git repo")
-	}
-}
-
-func TestInit_DryRunNoGit(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "dry-run-git")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	result, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "dry-run",
-		DryRun:     true,
-		NoRegister: true,
-	})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// Check that git was NOT initialized (dry run)
-	if result.GitInitialized {
-		t.Error("GitInitialized should be false in dry run mode")
-	}
-
-	// Check that .git directory was NOT created
-	gitDir := filepath.Join(campaignDir, ".git")
-	if _, err := os.Stat(gitDir); err == nil {
-		t.Errorf(".git directory should not exist in dry run mode")
-	}
-}
-
-func TestInit_DescriptionAndMission(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "described-campaign")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:        "described",
-		Description: "A well-described campaign",
-		Mission:     "Build something awesome",
-		NoRegister:  true,
-	})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// Load config and verify description and mission
-	cfg, err := config.LoadCampaignConfig(ctx, campaignDir)
-	if err != nil {
-		t.Fatalf("LoadCampaignConfig() error = %v", err)
-	}
-
-	if cfg.Description != "A well-described campaign" {
-		t.Errorf("Description = %q, want %q", cfg.Description, "A well-described campaign")
-	}
-	if cfg.Mission != "Build something awesome" {
-		t.Errorf("Mission = %q, want %q", cfg.Mission, "Build something awesome")
-	}
-}
-
-func TestInit_DefaultDescription(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "default-desc")
-	os.MkdirAll(campaignDir, 0755)
-
-	ctx := context.Background()
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "default-desc",
-		NoRegister: true,
-		// No description provided - should get default
-	})
-
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	cfg, err := config.LoadCampaignConfig(ctx, campaignDir)
-	if err != nil {
-		t.Fatalf("LoadCampaignConfig() error = %v", err)
-	}
-
-	expectedDesc := "Campaign: default-desc"
-	if cfg.Description != expectedDesc {
-		t.Errorf("Description = %q, want %q (default format)", cfg.Description, expectedDesc)
-	}
-}
-
-func TestInit_CreatesCanonicalIntentDirectories(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "canonical-intents")
-	if err := os.MkdirAll(campaignDir, 0755); err != nil {
-		t.Fatalf("failed to create campaign dir: %v", err)
-	}
-
-	ctx := context.Background()
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:        "canonical-intents",
-		NoRegister:  true,
-		SkipGitInit: true,
-	})
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	for _, relDir := range []string{
-		".campaign/intents/inbox",
-		".campaign/intents/ready",
-		".campaign/intents/active",
-		".campaign/intents/dungeon/done",
-		".campaign/intents/dungeon/killed",
-		".campaign/intents/dungeon/archived",
-		".campaign/intents/dungeon/someday",
-	} {
-		path := filepath.Join(campaignDir, filepath.FromSlash(relDir))
-		if info, statErr := os.Stat(path); statErr != nil {
-			t.Fatalf("expected canonical intent directory %s to exist: %v", relDir, statErr)
-		} else if !info.IsDir() {
-			t.Fatalf("expected %s to be a directory", relDir)
-		}
-	}
-}
-
-func TestInit_RepairPreservesMission(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "repair-mission")
-	os.MkdirAll(filepath.Join(campaignDir, config.CampaignDir), 0755)
-
-	ctx := context.Background()
-
-	// Create initial campaign with mission
-	initialCfg := &config.CampaignConfig{
-		ID:          "test-id",
-		Name:        "repair-mission",
-		Type:        config.CampaignTypeProduct,
-		Description: "Original description",
-		Mission:     "Original mission",
-		CreatedAt:   time.Now(),
-	}
-	if err := config.SaveCampaignConfig(ctx, campaignDir, initialCfg); err != nil {
-		t.Fatalf("SaveCampaignConfig() error = %v", err)
-	}
-
-	// Run repair without providing new mission
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "repair-mission",
-		Repair:     true,
-		NoRegister: true,
-	})
-
-	if err != nil {
-		t.Fatalf("Init() with repair error = %v", err)
-	}
-
-	// Load config and verify mission was preserved
-	cfg, err := config.LoadCampaignConfig(ctx, campaignDir)
-	if err != nil {
-		t.Fatalf("LoadCampaignConfig() error = %v", err)
-	}
-
-	if cfg.Mission != "Original mission" {
-		t.Errorf("Mission = %q, want %q (should preserve existing)", cfg.Mission, "Original mission")
-	}
-}
-
-func TestInit_RepairMigratesLegacyIntentState(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "repair-intents")
-	if err := os.MkdirAll(filepath.Join(campaignDir, config.CampaignDir), 0755); err != nil {
-		t.Fatalf("failed to create campaign dir: %v", err)
-	}
-
-	ctx := context.Background()
-	initialCfg := &config.CampaignConfig{
-		ID:          "repair-intents-id",
-		Name:        "repair-intents",
-		Type:        config.CampaignTypeProduct,
-		Description: "Repair intents",
-		CreatedAt:   time.Now(),
-	}
-	if err := config.SaveCampaignConfig(ctx, campaignDir, initialCfg); err != nil {
-		t.Fatalf("SaveCampaignConfig() error = %v", err)
-	}
-
-	legacyJumps := &config.JumpsConfig{
-		Paths: config.CampaignPaths{
-			Workflow: "workflow/",
-			Intents:  "workflow/intents/",
-		},
-		Shortcuts: map[string]config.ShortcutConfig{
-			"i": {
-				Path:        "workflow/intents/",
-				Concept:     "intent",
-				Description: "Legacy intents shortcut",
-				Source:      config.ShortcutSourceAuto,
-			},
-		},
-	}
-	if err := config.SaveJumpsConfig(ctx, campaignDir, legacyJumps); err != nil {
-		t.Fatalf("SaveJumpsConfig() error = %v", err)
-	}
-
-	legacyIntent := &intent.Intent{
-		ID:        "20260316-repair-intent",
-		Title:     "Repair Intent",
-		Status:    intent.StatusInbox,
-		Type:      intent.TypeFeature,
-		CreatedAt: time.Date(2026, 3, 16, 12, 0, 0, 0, time.UTC),
-		UpdatedAt: time.Date(2026, 3, 16, 12, 0, 0, 0, time.UTC),
-		Content:   "legacy intent\n",
-	}
-	legacyData, err := intent.SerializeIntent(legacyIntent)
-	if err != nil {
-		t.Fatalf("SerializeIntent() error = %v", err)
-	}
-
-	legacyIntentPath := filepath.Join(campaignDir, "workflow", "intents", "inbox", legacyIntent.ID+".md")
-	if err := os.MkdirAll(filepath.Dir(legacyIntentPath), 0755); err != nil {
-		t.Fatalf("failed to create legacy intent dir: %v", err)
-	}
-	if err := os.WriteFile(legacyIntentPath, legacyData, 0644); err != nil {
-		t.Fatalf("failed to write legacy intent: %v", err)
-	}
-	legacyObeyPath := filepath.Join(campaignDir, "workflow", "intents", "OBEY.md")
-	legacyObey := "# legacy intent docs\n"
-	if err := os.WriteFile(legacyObeyPath, []byte(legacyObey), 0644); err != nil {
-		t.Fatalf("failed to write legacy intent OBEY.md: %v", err)
-	}
-	legacyAuditPath := filepath.Join(campaignDir, "workflow", "intents", ".intents.jsonl")
-	if err := os.WriteFile(legacyAuditPath, []byte("{\"event\":\"create\"}\n"), 0644); err != nil {
-		t.Fatalf("failed to write legacy audit log: %v", err)
-	}
-
-	plan, err := ComputeRepairPlan(ctx, campaignDir, InitOptions{
-		Name:       "repair-intents",
-		Type:       config.CampaignTypeProduct,
-		Repair:     true,
-		NoRegister: true,
-	})
-	if err != nil {
-		t.Fatalf("ComputeRepairPlan() error = %v", err)
-	}
-
-	_, err = Init(ctx, campaignDir, InitOptions{
-		Name:        "repair-intents",
-		Type:        config.CampaignTypeProduct,
-		Repair:      true,
-		RepairPlan:  plan,
-		NoRegister:  true,
-		SkipGitInit: true,
-	})
-	if err != nil {
-		t.Fatalf("Init() with repair error = %v", err)
-	}
-
-	canonicalIntentPath := filepath.Join(campaignDir, ".campaign", "intents", "inbox", legacyIntent.ID+".md")
-	if _, err := os.Stat(canonicalIntentPath); err != nil {
-		t.Fatalf("expected canonical intent at %s: %v", canonicalIntentPath, err)
-	}
-	if _, err := os.Stat(legacyIntentPath); !os.IsNotExist(err) {
-		t.Fatalf("legacy intent should be removed after repair, err = %v", err)
-	}
-
-	canonicalAuditPath := filepath.Join(campaignDir, ".campaign", "intents", ".intents.jsonl")
-	if _, err := os.Stat(canonicalAuditPath); err != nil {
-		t.Fatalf("expected canonical audit log at %s: %v", canonicalAuditPath, err)
-	}
-	canonicalObeyPath := filepath.Join(campaignDir, ".campaign", "intents", "OBEY.md")
-	canonicalObey, err := os.ReadFile(canonicalObeyPath)
-	if err != nil {
-		t.Fatalf("expected canonical intent marker at %s: %v", canonicalObeyPath, err)
-	}
-	if string(canonicalObey) != legacyObey {
-		t.Fatalf("canonical intent marker = %q, want %q", string(canonicalObey), legacyObey)
-	}
-	if _, err := os.Stat(legacyAuditPath); !os.IsNotExist(err) {
-		t.Fatalf("legacy audit log should be removed after repair, err = %v", err)
-	}
-	if _, err := os.Stat(legacyObeyPath); !os.IsNotExist(err) {
-		t.Fatalf("legacy intent scaffold docs should be removed after repair, err = %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(campaignDir, "workflow", "intents")); !os.IsNotExist(err) {
-		t.Fatalf("legacy workflow/intents tree should be removed after repair, err = %v", err)
-	}
-}
-
-func TestInit_RepairUpdatesMission(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "repair-update")
-	os.MkdirAll(filepath.Join(campaignDir, config.CampaignDir), 0755)
-
-	ctx := context.Background()
-
-	// Create initial campaign with mission
-	initialCfg := &config.CampaignConfig{
-		ID:          "test-id",
-		Name:        "repair-update",
-		Type:        config.CampaignTypeProduct,
-		Description: "Original description",
-		Mission:     "Original mission",
-		CreatedAt:   time.Now(),
-	}
-	if err := config.SaveCampaignConfig(ctx, campaignDir, initialCfg); err != nil {
-		t.Fatalf("SaveCampaignConfig() error = %v", err)
-	}
-
-	// Run repair with new mission
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "repair-update",
-		Mission:    "Updated mission via repair",
-		Repair:     true,
-		NoRegister: true,
-	})
-
-	if err != nil {
-		t.Fatalf("Init() with repair error = %v", err)
-	}
-
-	// Load config and verify mission was updated
-	cfg, err := config.LoadCampaignConfig(ctx, campaignDir)
-	if err != nil {
-		t.Fatalf("LoadCampaignConfig() error = %v", err)
-	}
-
-	if cfg.Mission != "Updated mission via repair" {
-		t.Errorf("Mission = %q, want %q (should update)", cfg.Mission, "Updated mission via repair")
-	}
-}
-
-func TestInit_RepairPreservesLegacyConceptList(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "repair-legacy-concepts")
-	if err := os.MkdirAll(filepath.Join(campaignDir, config.CampaignDir), 0755); err != nil {
-		t.Fatalf("failed to create campaign dir: %v", err)
-	}
-
-	ctx := context.Background()
-
-	initialCfg := &config.CampaignConfig{
-		ID:          "legacy-id",
-		Name:        "repair-legacy-concepts",
-		Type:        config.CampaignTypeProduct,
-		Description: "Legacy concept config",
-		Mission:     "Keep legacy concepts",
-		CreatedAt:   time.Now(),
-		ConceptList: []config.ConceptEntry{
-			{Name: "design", Path: "workflow/design/", Description: "Design"},
-			{Name: "dungeon", Path: "dungeon/", Description: "Legacy dungeon concept"},
-		},
-	}
-	if err := config.SaveCampaignConfig(ctx, campaignDir, initialCfg); err != nil {
-		t.Fatalf("SaveCampaignConfig() error = %v", err)
-	}
-
-	_, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "repair-legacy-concepts",
-		Repair:     true,
-		NoRegister: true,
-	})
-	if err != nil {
-		t.Fatalf("Init() with repair error = %v", err)
-	}
-
-	cfg, err := config.LoadCampaignConfig(ctx, campaignDir)
-	if err != nil {
-		t.Fatalf("LoadCampaignConfig() error = %v", err)
-	}
-
-	if len(cfg.ConceptList) != 2 {
-		t.Fatalf("len(ConceptList) = %d, want 2", len(cfg.ConceptList))
-	}
-
-	var hasDungeon bool
-	for _, c := range cfg.ConceptList {
-		if c.Name == "dungeon" && c.Path == "dungeon/" {
-			hasDungeon = true
-			break
-		}
-	}
-	if !hasDungeon {
-		t.Fatal("repair should preserve explicit legacy dungeon concept entry")
-	}
-}
-
-func TestScaffoldInit_SkipFestPassThrough(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "skip-fest-passthrough")
-	if err := os.MkdirAll(campaignDir, 0755); err != nil {
-		t.Fatalf("failed to create campaign dir: %v", err)
-	}
-
-	ctx := context.Background()
-	opts := InitOptions{
-		Name:        "skip-fest-passthrough",
-		NoRegister:  true,
-		SkipGitInit: true,
-		SkipFest:    true,
-	}
-
-	result, err := Init(ctx, campaignDir, opts)
-	if err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	// scaffold.Init does not produce festivals/ on its own (festival init is cmd-layer owned).
-	for _, d := range result.DirsCreated {
-		if strings.HasSuffix(d, "festivals") || strings.Contains(d, "festivals/") {
-			t.Errorf("scaffold.Init should not create festivals/ directory; got DirsCreated entry: %s", d)
-		}
-	}
-	// Result should have no error.
-	if result.CampaignRoot != campaignDir {
-		t.Errorf("CampaignRoot = %q, want %q", result.CampaignRoot, campaignDir)
-	}
-	// SkipFest field is preserved on opts (sanity check that it was not mutated).
-	if !opts.SkipFest {
-		t.Error("opts.SkipFest should remain true after Init()")
-	}
-}
-
-func TestInit_RepairRestoresQuestScaffold(t *testing.T) {
-	tmpDir := t.TempDir()
-	tmpDir, _ = filepath.EvalSymlinks(tmpDir)
-	t.Setenv("XDG_CONFIG_HOME", tmpDir)
-
-	campaignDir := filepath.Join(tmpDir, "repair-quest")
-	if err := os.MkdirAll(campaignDir, 0755); err != nil {
-		t.Fatalf("failed to create campaign dir: %v", err)
-	}
-
-	ctx := context.Background()
-	if _, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "repair-quest",
-		NoRegister: true,
-	}); err != nil {
-		t.Fatalf("Init() error = %v", err)
-	}
-
-	removed := []string{
-		filepath.Join(campaignDir, quest.RootDirName, quest.DefaultDirName, quest.FileName),
-		filepath.Join(campaignDir, quest.RootDirName, "dungeon", "OBEY.md"),
-	}
-	for _, path := range removed {
-		if err := os.Remove(path); err != nil {
-			t.Fatalf("os.Remove(%q) error = %v", path, err)
-		}
-	}
-
-	if _, err := Init(ctx, campaignDir, InitOptions{
-		Name:       "repair-quest",
-		Repair:     true,
-		NoRegister: true,
-	}); err != nil {
-		t.Fatalf("Init() with repair error = %v", err)
-	}
-
-	for _, path := range removed {
-		if _, err := os.Stat(path); os.IsNotExist(err) {
-			t.Errorf("expected repaired quest scaffold path to exist: %s", path)
-		}
-	}
-
-	// Verify .active file is NOT recreated on repair
-	activePath := filepath.Join(campaignDir, quest.RootDirName, ".active")
-	if _, err := os.Stat(activePath); !os.IsNotExist(err) {
-		t.Errorf(".active file should not be created on repair: %s", activePath)
 	}
 }
